refactor: organize agents by name, read prompts from markdown files
Eliminates duplication between prompt files and inlined systemPrompt
strings. Each agent is now a folder with a prompt.md (source of truth)
and a definition.ts that reads from it at load time.
