+++ libs/lib-agent/src/lib/agents/library-assistant/prompt.md
